# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TodoList API",
    "version": "1.0.0",
    "description": "TodoList Phase 1 REST API specification derived from docs. The API supports email/password authentication, JWT-protected Todo and Category management, filtering, overdue detection, and profile lookup."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/todos": {
      "get": {
        "tags": [
          "Todos"
        ],
        "summary": "Todo 목록 조회",
        "description": "인증 사용자 본인의 Todo만 조회한다. 카테고리, 상태, 미분류, 기한 초과 조건으로 필터링할 수 있다. 기본 정렬은 문서의 미결정 사항이며 구현 시 확정한다.",
        "operationId": "getTodos",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "미지정 시 전체 상태를 조회한다.",
            "schema": {
              "$ref": "#/components/schemas/TodoStatus"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "description": "특정 카테고리의 Todo만 조회한다. uncategorized=true와 동시에 사용할 수 없다.",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "uncategorized",
            "in": "query",
            "required": false,
            "description": "true면 categoryId가 null인 Todo만 조회한다.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "overdue",
            "in": "query",
            "required": false,
            "description": "true면 dueDate가 서버 기준 오늘보다 이전이고 status가 in_progress인 Todo만 조회한다. status=done과 동시에 사용하면 400을 반환한다.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "본인 Todo 목록 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Todos"
        ],
        "summary": "Todo 생성",
        "description": "새 Todo를 생성한다. 생성 시 status는 in_progress이고 completedAt은 null이다. categoryId는 선택값이며 본인 소유 카테고리여야 한다.",
        "operationId": "createTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoRequest"
              },
              "examples": {
                "withCategory": {
                  "value": {
                    "title": "주간 보고서 작성",
                    "description": "금요일 회의 전까지 초안 작성",
                    "dueDate": "2026-04-29",
                    "categoryId": 1
                  }
                },
                "withoutCategory": {
                  "value": {
                    "title": "헬스장 등록",
                    "description": null,
                    "dueDate": null,
                    "categoryId": null
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성된 Todo 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/todos/{todoId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TodoId"
        }
      ],
      "get": {
        "tags": [
          "Todos"
        ],
        "summary": "Todo 상세 조회",
        "description": "인증 사용자 본인의 Todo 상세 정보를 조회한다.",
        "operationId": "getTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Todo 상세 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "tags": [
          "Todos"
        ],
        "summary": "Todo 내용 수정",
        "description": "Todo의 제목, 설명, 종료일, 카테고리를 수정한다. status와 completedAt은 이 API에서 직접 수정하지 않는다.",
        "operationId": "updateTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 Todo 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Todos"
        ],
        "summary": "Todo 삭제",
        "description": "인증 사용자 본인의 Todo를 삭제한다.",
        "operationId": "deleteTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptySuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/todos/{todoId}/status": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TodoId"
        }
      ],
      "patch": {
        "tags": [
          "Todos"
        ],
        "summary": "Todo 상태 변경",
        "description": "Todo 상태를 in_progress 또는 done으로 변경한다. done으로 변경하면 서버 현재 시각을 completedAt에 기록하고, in_progress로 변경하면 completedAt을 null로 초기화한다.",
        "operationId": "updateTodoStatus",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoStatusRequest"
              },
              "examples": {
                "complete": {
                  "value": {
                    "status": "done"
                  }
                },
                "reopen": {
                  "value": {
                    "status": "in_progress"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "상태 변경된 Todo 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "카테고리 목록 조회",
        "description": "인증 사용자 본인의 카테고리 목록만 조회한다.",
        "operationId": "getCategories",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "본인 카테고리 목록 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "카테고리 생성",
        "description": "카테고리를 생성한다. 이름은 사용자 내에서 고유해야 하며 사용자당 최대 20개까지 생성할 수 있다.",
        "operationId": "createCategory",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "name": "업무"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성된 카테고리 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/DuplicateCategoryName"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/categories/{categoryId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CategoryId"
        }
      ],
      "patch": {
        "tags": [
          "Categories"
        ],
        "summary": "카테고리 이름 수정",
        "description": "인증 사용자 본인의 카테고리 이름을 수정한다.",
        "operationId": "updateCategory",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "name": "개인"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 카테고리 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/DuplicateCategoryName"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "카테고리 삭제",
        "description": "인증 사용자 본인의 카테고리를 삭제한다. 해당 카테고리에 속한 Todo는 삭제하지 않고 categoryId를 null로 변경하여 미분류 처리한다.",
        "operationId": "deleteCategory",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "카테고리 삭제 성공 및 미분류 전환 결과 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCategoryResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT HS-512 access token issued by POST /api/auth/login."
      }
    },
    "parameters": {
      "TodoId": {
        "name": "todoId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        },
        "description": "Todo ID"
      },
      "CategoryId": {
        "name": "categoryId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        },
        "description": "Category ID"
      }
    },
    "responses": {
      "ValidationError": {
        "description": "입력값 검증 실패",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validation": {
                "value": {
                  "success": false,
                  "message": "입력값이 올바르지 않습니다.",
                  "code": "VALIDATION_ERROR"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "인증 토큰 없음, 위조 또는 만료",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "success": false,
                  "message": "인증 토큰이 필요합니다.",
                  "code": "UNAUTHORIZED"
                }
              }
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "요청 리소스에 대한 권한 없음",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "forbidden": {
                "value": {
                  "success": false,
                  "message": "권한이 없습니다.",
                  "code": "FORBIDDEN"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "리소스를 찾을 수 없음",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "notFound": {
                "value": {
                  "success": false,
                  "message": "리소스를 찾을 수 없습니다.",
                  "code": "NOT_FOUND"
                }
              }
            }
          }
        }
      },
...
      "DuplicateCategoryName": {
        "description": "동일 사용자 내 카테고리 이름 중복",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "duplicateCategory": {
                "value": {
                  "success": false,
                  "message": "이미 존재하는 카테고리 이름입니다.",
                  "code": "DUPLICATE_CATEGORY_NAME"
                }
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "서버 오류",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "internal": {
                "value": {
                  "success": false,
                  "message": "서버 오류가 발생했습니다.",
                  "code": "INTERNAL_ERROR"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "TodoStatus": {
        "type": "string",
        "enum": [
          "in_progress",
          "done"
        ],
        "description": "in_progress=진행 중, done=완료"
      },
      "CategorySummary": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "maxLength": 20,
            "example": "업무"
          }
        }
      },
      "Todo": {
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "isOverdue",
          "createdAt"
        ],
"properties": {
            "id": {
              "type": "integer",
              "format": "int64",
              "example": 1
            },
            "userId": {
              "type": "integer",
              "format": "int64",
              "example": 1
            },
            "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "주간 보고서 작성"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "금요일 회의 전까지 초안 작성"
          },
          "status": {
            "$ref": "#/components/schemas/TodoStatus"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-04-30"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "example": 1
          },
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CategorySummary"
              }
            ],
            "nullable": true
          },
          "isOverdue": {
            "type": "boolean",
            "description": "true when dueDate is before the server 기준 오늘 and status is in_progress.",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-28T09:00:00Z"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": null
          }
        }
      },
      "CreateTodoRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "주간 보고서 작성"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "금요일 회의 전까지 초안 작성"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-04-29"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "minimum": 1,
            "example": 1
          }
        }
      },
      "UpdateTodoRequest": {
        "type": "object",
        "minProperties": 1,
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "분기 보고서 제출"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "수정된 상세 설명"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-04-30"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "minimum": 1,
            "example": 1
          }
        }
      },
      "UpdateTodoStatusRequest": {
        "type": "object",
        "required": [
          "status"
        ],
        "additionalProperties": false,
        "properties": {
          "status": {
            "$ref": "#/components/schemas/TodoStatus"
          }
        }
      },
      "TodoResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/Todo"
          }
        }
      },
      "TodoListResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Todo"
            }
          }
        }
      },
      "EmptySuccessResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "nullable": true,
            "example": null,
            "description": "삭제처럼 별도 반환 데이터가 없는 성공 응답"
          }
        }
      },
      "Category": {
        "type": "object",
        "required": [
          "id",
          "name",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20,
            "example": "업무"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-28T09:00:00Z"
          }
        }
      },
      "CategoryRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20,
            "example": "업무"
          }
        }
      },
      "CategoryResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/Category"
          }
        }
      },
      "CategoryListResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          }
        }
      },
      "DeleteCategoryResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "required": [
              "deletedCategoryId",
              "uncategorizedTodoCount"
            ],
            "properties": {
              "deletedCategoryId": {
                "type": "integer",
                "format": "int64",
                "example": 4
              },
              "uncategorizedTodoCount": {
                "type": "integer",
                "example": 2,
                "description": "삭제된 카테고리에서 미분류로 전환된 Todo 수"
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "code"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "입력값이 올바르지 않습니다."
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "DUPLICATE_EMAIL",
              "DUPLICATE_CATEGORY_NAME",
              "CATEGORY_LIMIT_EXCEEDED",
              "INVALID_CREDENTIALS",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "INTERNAL_ERROR"
            ],
            "example": "VALIDATION_ERROR"
          }
        }
      }
    }
  }
}